Fix test failure- Unrestricted schema compatibility checks
Fix test failure- Schema compatibility changes ISPN-15974

Fix test failure- Schema compatibility changes ISPN-15974
