+++ data-index/data-index-service/data-index-service-infinispan/src/test/docker/docker-compose.yml
@@ -55,7 +55,7 @@ services:
       LOG_DIR: "/tmp/logs"
 
   infinispan:
-    image: infinispan/server:14.0.4.Final
+    image: infinispan/server:15.0.8.Final
     container_name: data-index-infinispan
     ports:
       - "11222:11222"
