+++ data-index/data-index-service/data-index-service-mongodb/src/test/java/org/kie/kogito/index/service/MongoDomainIndexingServiceIT.java
@@ -18,11 +18,18 @@
  */
 package org.kie.kogito.index.service;
 
+import io.restassured.http.ContentType;
+import org.junit.jupiter.api.Test;
 import org.kie.kogito.index.service.test.InMemoryMessageTestProfile;
 import org.kie.kogito.index.test.TestUtils;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
+
+import java.util.Collection;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.CoreMatchers.isA;
 
 @QuarkusTest
 @TestProfile(InMemoryMessageTestProfile.class)
@@ -36,5 +43,58 @@ class MongoDomainIndexingServiceIT extends AbstractDomainIndexingServiceIT {
     @Override
     protected String getUserTaskProtobufFileContent() throws Exception {
         return TestUtils.readFileContent("deals-mongo.proto");
+    }
+    @Test
+    void testAddProtoFileTwice() throws Exception {
+        protobufService.registerProtoBufferType(getProtoBufferFileV1());
+        given().contentType(ContentType.JSON)
+                .body("{ \"query\" : \"{Game{ player, id, name, metadata { processInstances { id } } } }\" }")
+                .when().post("/graphql")
+                .then().log().ifValidationFails().statusCode(200).body("data.Game", isA(Collection.class));
+        given().contentType(ContentType.JSON)
+                .body("{ \"query\" : \"{ProcessInstances{ id, processId, rootProcessId, rootProcessInstanceId, parentProcessInstanceId } }\" }")
+                .when().post("/graphql")
+                .then().log().ifValidationFails().statusCode(200).body("data.ProcessInstances", isA(Collection.class));
+
+        protobufService.registerProtoBufferType(getProtoBufferFileV2());
+        given().contentType(ContentType.JSON)
+                .body("{ \"query\" : \"{Game{ id, name, company, metadata { processInstances { id } } } }\" }")
+                .when().post("/graphql")
+                .then().log().ifValidationFails().statusCode(200).body("data.Game", isA(Collection.class));
+        given().contentType(ContentType.JSON)
+                .body("{ \"query\" : \"{ProcessInstances{ id, processId, rootProcessId, rootProcessInstanceId, parentProcessInstanceId } }\" }")
+                .when().post("/graphql")
+                .then().log().ifValidationFails().statusCode(200).body("data.ProcessInstances", isA(Collection.class));
+    }
+    private String getProtoBufferFileV1() {
+        return "package org.demo;\n" +
+                "import \"kogito-index.proto\";\n" +
+                "option kogito_model=\"Game\";\n" +
+                "option kogito_id=\"game\";\n" +
+                "/* @Indexed */\n" +
+                "message Game {\n" +
+                "   optional string player = 1;\n" +
+                "   /* @Field(index = Index.YES, store = Store.YES) @SortableField */\n" +
+                "   optional string id = 2;\n" +
+                "   optional string name = 3;\n" +
+                "   optional org.kie.kogito.index.model.KogitoMetadata metadata = 4;\n" +
+                "}\n" +
+                "\n";
+    }
+
+    private String getProtoBufferFileV2() {
+        return "package org.demo;\n" +
+                "import \"kogito-index.proto\";\n" +
+                "option kogito_model=\"Game\";\n" +
+                "option kogito_id=\"game\";\n" +
+                "/* @Indexed */\n" +
+                "message Game {\n" +
+                "   /* @Field(index = Index.YES, store = Store.YES) @SortableField */\n" +
+                "   optional string id = 1;\n" +
+                "   optional string name = 2;\n" +
+                "   optional string company = 3;\n" +
+                "   optional org.kie.kogito.index.model.KogitoMetadata metadata = 4;\n" +
+                "}\n" +
+                "\n";
     }
 }
